+++ packages/apptive_grid_core/test/currency_test.dart
@@ -341,6 +341,7 @@ void main() {
           },
           schema: {"type": "number", "format": "float"},
         ),
+        type: 'multiSelectDropdown',
       );
 
       expect(fromJson.cast<CurrencyDataEntity>(), equals(direct));
@@ -355,6 +356,7 @@ void main() {
         property: 'New field',
         data: directEntity,
         field: field,
+        type: 'multiSelectDropdown',
       );
 
       expect(
@@ -365,6 +367,7 @@ void main() {
           component.data,
           component.options,
           component.required,
+          component.type,
         ),
       );
     });