+++ packages/apptive_grid_core/CHANGELOG.md
@@ -1,7 +1,8 @@
 ## 0.14.0
 **BREAKING CHANGE**
-- Restructure `ApptiveGridAuthenticator` to handle changes in options correctly by using a specific `ApptiveGridClient` inside the authenticator
-- Update dependencies
+* Restructure `ApptiveGridAuthenticator` to handle changes in options correctly by using a specific `ApptiveGridClient` inside the authenticator
+* Update dependencies
+* Use new Website for authentication
 
 ## 0.13.4+1
 * Fix race condition when refreshing the access token